+++ source/added_value/items_list_directive.py
@@ -11,11 +11,10 @@ from docutils.parsers.rst.states import Body
 from docutils.parsers.rst import Directive, directives
 from docutils.parsers.rst.directives import unchanged_required, unchanged
 from six import StringIO
-from sphinx.ext.autosummary import import_by_name
 
-from added_value import common_options
 from added_value.common_options import NAME_OPTION, CLASS_OPTION
 from added_value.grammatical_conjunctions import list_conjunction
+from added_value.importer import import_object
 from added_value.multisort import asc, dec, as_is
 from added_value.non_string_iterable import NonStringIterable
 
@@ -473,13 +472,7 @@ class ItemsListDirective(Directive):
     def run(self):
 
         obj_name = self.arguments[0]
-
-        try:
-            prefixed_name, obj, parent, modname = import_by_name(obj_name)
-        except ImportError:
-            raise self.error(
-                "Could not locate Python object {} ({} directive).".format(obj_name, self.name)
-            )
+        obj, prefixed_name = import_object(obj_name, context=self)
 
         list_types = self.list_types