+++ source/added_value/pyobj_role.py
@@ -1,6 +1,6 @@
 from functools import partial
 
-from sphinx.ext.autosummary import import_by_name
+from sphinx.ext.autosummary import import_by_name, ImportExceptionGroup
 
 
 def make_pyobj_role(make_node):
@@ -31,7 +31,7 @@ def pyobj_role(make_node, name, rawtext, text, lineno, inliner, options=None, co
 
     try:
         prefixed_name, obj, parent, modname = import_by_name(text)
-    except ImportError:
+    except (ImportError, ImportExceptionGroup):
         msg = inliner.reporter.error("Could not locate Python object {}".format(text), line=lineno)
         prb = inliner.problematic(rawtext, rawtext, msg)
         return [prb], [msg]