+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+$(next_version) - Updates copyright notices.
+
 0.22.0 - Minor fixes to the documentation.
 
 0.21.5 - Build system updates